% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Identical_samples.R
\name{AddGroup2SampleInfo}
\alias{AddGroup2SampleInfo}
\title{Adding information about group of identical samples to a Genotype object}
\usage{
AddGroup2SampleInfo(object, group = NULL)
}
\arguments{
\item{id_table}{\code{Genotype}. The object to add the information to.}

\item{\code{list}.}{A list of \code{character} vectors. Each one is a set
of sample names which are all genetically identical}
}
\value{
\code{Genotype}. The object with the sample information containing
the identity group.
}
\description{
After applying the functions \code{FindIdentical} and \code{IdenticalGroups},
this function allows you to add an "Identity_group" column to the object's
\code{SampleInfo}, which specifies for samples that are a part of genetically
identical group, the group they belong to.
}
\examples{
data("Mango_row")
F0.05 = FindIdentical(MxS, error.freq=0.05)
group = IdenticalGroups(F0.05)
names(group) = paste("Group", names(group))
MxS = AddGroup2SampleInfo(MxS, group)
head(getSampleInfo(MxS))
getSampleInfo(MxS)[group[[7]], ]
}
\seealso{
\code{\link{FindIdentical}} for finding pairs of identical samples,
\code{\link{IdenticalGroups}} for creating groups of identical samples, and
\code{\link{Unique}} for filtering all samples but one from each identity
group.
}

