% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cervus_interface.R
\name{addSampleInfo2cervus}
\alias{addSampleInfo2cervus}
\title{Adding sample information to the cervus result table}
\usage{
addSampleInfo2cervus(cervus, SampleInfo, add.columns)
}
\arguments{
\item{cervus}{\code{data.frame}. The cervus results table, obtained with
\code{\link{read.cervus}}.}

\item{SampleInfo}{\code{SampleInfo}. Sample information table}

\item{add.columns}{\code{character}. Names of columns in \code{SampleInfo}
to add to the Cervus results table}
}
\value{
\code{data.frame}. The cervus results table including the new
columns.
}
\description{
Adding sample information to the cervus result table
}

