% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{AlleleCounting}
\alias{AlleleCounting}
\title{Counting genotype frequencies}
\usage{
AlleleCounting(object)
}
\arguments{
\item{object.}{\code{Genotype}. The object to count the genotypes in.}
}
\value{
\code{AlleleStat}. An object contianing the matrix of genotype
 frequencies. Each row corresponds to a marker, and the columns list the
 number of individual samples with "XX", "XY" and "YY" genotypes.
}
\description{
Counting the frequency of each genotype, per each marker
}
\examples{
data("Mango_raw")
MarkersStat = AlleleCounting(MxS)
GF = genotype_freq(MarkersStat)
head(GF)
}

