% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\docType{class}
\name{AlleleStat-class}
\alias{AlleleStat-class}
\title{A S4 class for a matrix genotype frequencies.}
\description{
A S4 class for a matrix genotype frequencies.
}
\section{Slots}{

\describe{
\item{\code{genotype_freq}}{\code{matrix}: A matrix of genotype frequencies. Each
row corresponds to a marker, and the columns list the number of individual
samples with "XX", "XY" and "YY" genotypes.}
}}

