% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust_tree.R
\name{cluster_STRUCTURE_plot}
\alias{cluster_STRUCTURE_plot}
\title{Drows a population \code{phylo} 'tree' with STRUCTRE results at the bottom}
\usage{
cluster_STRUCTURE_plot(tree, anc, addlegend = T, parameter_list = list())
}
\arguments{
\item{tree}{\code{phylo}. The tree to plot}

\item{anc}{\code{data.frame}. A table with the summary of STRUCTURE results.
For STRUCURE with K populations, this table should contain K columns, named
"cluster_1", "cluster_2", ... "cluster_K", specifying the relative
ancestries of each individual. The number of rows should be identical to the
number of individuals, and it has to be ordered according to the order of
tips in the tree.}

\item{addlegend}{\code{logical}. Should a legend be added?}

\item{parameter_list}{\code{list}. A named list of additional arguments to
pass to the function. possible arguments are:
\describe{
  \item{"plotname"}{\code{character}. The tiff image file name. default is
"cluster_STRUCTURE".}
  \item{"plotname_prefix"}{\code{character}. A prefix for the tiff image
  file name. default is NULL.}
  \item{"BarplotColour"}{The colours to use for the STRUCTURE results bar
  plot.}
  \item{"SampleInfo"}{\code{SampleInfo}. A table of sample information. The
   row names of the table should be identical to the \code{object$tip.label}
   .}
  \item{"colourby"}{\code{character}. Colour the tree tip lables according
  to this column in \code{SampleInfo}.}
  \item{"ColourPalette"}{\code{character}. ColourPalette for the tree tip
lables.}
  \item{"tip.label.by"}{\code{character}. Which column in SampleInfo to
  colour the tree tip labels by.}
  \item{"legend.x"}{\code{numeric}. legend x position}
  \item{"legend.y"}{\code{numeric}. legend y position}
  \item{"legend.cex"}{\code{numeric}. legend font size}
  \item{"mark.node.labels"}{\code{logical}. Should the nodes be labeled?}
  \item{"node.label.cutoff"}{\code{numeric}. What is the minimum value of
  node label to include in the tree graph? The node labels numeric values
  specifying the reproducibility of each node.}
}}
}
\value{
\code{NULL}.
}
\description{
This function creates a tiff image file, with a dendogram of the samples,
and STRUCTURE results below it.
}
\examples{
data("Avocado_phylo")
data("Avocado_clean123")
data("Avocado_ancestry")
SampleInfo = getSampleInfo(MxS)
cluster_STRUCTURE_plot(b.tree, anc=ancestry, parameter_list=list(SampleInfo=
getSampleInfo(MxS), colourby="Category", mark.node.labels=T))
}
\seealso{
\code{\link{bootstrap}} to create the tree, \code{\link{invert_tree}}
to change the leaf order of the tree, \code{\link{DrowPopulationTree}} to
plot only the tree, and \code{\link{getTreeLeafOrder}} to get the tree leaf
order.
}

