% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust_tree.R
\name{DrowPopulationTree}
\alias{DrowPopulationTree}
\title{Drows a population \code{phylo} 'tree' with sample info specifications}
\usage{
DrowPopulationTree(object, SampleInfo = NULL, tip.label.by = "Cultivar",
  colourby = NULL, ColourPalette = NULL, legend = TRUE, legend.x = NULL,
  legend.y = NULL, legend.cex = 1, mark.node.labels = FALSE)
}
\arguments{
\item{object}{\code{phylo}. The tree to plot}

\item{SampleInfo}{\code{SampleInfo}. A table of sample information. The row
names of the table should be identical to the \code{object$tip.label}.}

\item{tip.label.by}{\code{character}. Which column in SampleInfo to colour
the tree tip labels by.}

\item{colourby}{\code{character}. Colour tip lables according to this column
in \code{SampleInfo}.}

\item{ColourPalette}{\code{character}. ColourPalette for tip lables.}

\item{legend}{\code{logical}. Should a legend be added?}

\item{legend.x}{\code{numeric}. legend x position}

\item{legend.y}{\code{numeric}. legend y position}

\item{legend.cex}{\code{numeric}. legend font size}

\item{mark.node.labels}{\code{logical}. Should the nodes be labeled?}
}
\value{
\code{NULL}.
}
\description{
This function allows you to set the leave names and to colour the leaves
according to any category in the \code{SampleInfo}.
}
\examples{
data("Avocado_phylo")
data("Avocado_clean123")
SampleInfo = getSampleInfo(MxS)
DrowPopulationTree(b.tree, SampleInfo, colourby="Category", legend=TRUE)
}
\seealso{
\code{\link{bootstrap}} to create the tree, \code{\link{invert_tree}} to
 change the leaf order of the tree and  \code{\link{getTreeLeafOrder}} to get
 the tree leaf order.
}

