% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cervus_interface.R
\name{filter.cervus.results}
\alias{filter.cervus.results}
\title{Filtering 'Cervus' results}
\usage{
filter.cervus.results(cervus, mm_cutoff = Inf, lod_cutoff = 0)
}
\arguments{
\item{mm_cutoff}{\code{numeric}. The mis-match cutoff. The number of non-
mathcing markers that is allowed for a candidate parents-offspring trio.}

\item{lod_cutoff}{\code{numeric}. The lod score cutoff. A single numeric
value. The minimum lod score accepted for a candidate parents-offspring
trio. This value can be obtained from a simulation run of Cervus.}
}
\value{
\code{NULL}.
}
\description{
Cervus is a computer program for assignment of parents to their offspring
using genetic markers. This function allows you to filter its results and
leave only most probable candidate parents.
}
\examples{
data("cervus")
cervus = filter.cervus.results(cervus, mm_cutoff=9, lod_cutoff=9.84)

}

