% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.R
\name{FilterLDMarkers}
\alias{FilterLDMarkers}
\title{Eliminating linkage-disequilibrium redundancy among markers}
\usage{
FilterLDMarkers(object, by = "R^2", cutoff = NULL, score = NULL,
  print.n.markers = T)
}
\arguments{
\item{object}{\code{Genotype}. The object to be filtered}

\item{by}{\code{character}. The measure to determine if two markers are
linked. One of the followings:
\describe{
  \item{"R^2"}{Filter by r-square between markres}
  \item{"p" or "P"}{Filter by Chi-square p-value for marker independence}
}}

\item{cutoff}{\code{numberic} - The cutoff for filtering}

\item{score}{\code{numeric}. An optional score value for each marker, so
when choosing which marker to filter from a pair of linked markers, the one
with the higher score will be filtered, and the lower score will be kept.}
}
\value{
The same object with the filtered markers, so that all markers are
unlinked.
}
\description{
Leaving only one marker from each pair of linked markers. Using
\code{genetics} package function \code{LD} to calculate linkage
disequilibrium among marker.
}
\examples{
data("Genotype_43X80")
MarkerNo(MxS)
MxS = FilterLDMarkers(MxS)
}

