% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.R
\name{FilterMarkerByInfo}
\alias{FilterMarkerByInfo}
\title{Filtering markers by some \code{MarkerInfo} column}
\usage{
FilterMarkerByInfo(object, property.name = "", property.criterion = NULL,
  remove = "greaterOrEquall", print.n.markers = T)
}
\arguments{
\item{object}{\code{Genotype}. The object to be filtered.}

\item{property.name}{\code{character}. The name of the \code{MarkerInfo}
column to use for filtering}

\item{property.criterion}{a single \code{character} or \code{numeric}
variable. If \code{MarkerInfo$property.name} is a \code{character}, use a
\code{character} \code{property.criterion} to filter all markers with
\code{property.name} equal to \code{property.criterion} or filter all
markers with \code{property.name} different from \code{property.criterion},
depending on \code{remove}. If \code{MarkerInfo$property.name} is a
\code{numeric} vector, use a \code{numeric} \code{property.criterion} as a
cutoff.}

\item{remove}{\code{character}. One of the following:
    "equal", "different", "greater", "greaterOrEquall", "smaller" or
"smallerOrEquall".
Which markers to remove with respect to \code{property.criterion}}

\item{print.n.markers}{logical. Should the number of remaining markers be
printed on the screen?}
}
\value{
a filtered \code{Genotype} object, with updated filter table and
\code{MarkerInfo$Filtered_at_step}
}
\description{
This functions allows you to filter markers by any property, such as the
Fluidigm team rank, or whether this marker is mapped to the reference genome.
To use it, the relevant property should appear in the object's
\code{MarkerInfo} table.
}
\examples{
filter markers which Fluidigm team ranked >1
data("Mango_raw")
MarkerNo(MxS)
MxS = FilterMarkerByInfo(MxS, property.name="Category",
property.criterion=1, remove="greater")

leave only markers from chromosome 1
data("citrus_clean")
MarkerNo(MxS)
MxS = FilterMarkerByInfo(MxS, property.name="Genome_Chromosome",
property.criterion="chr1", remove="different")
}

