% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.R
\name{FilterMarkerByMajorGenotype}
\alias{FilterMarkerByMajorGenotype}
\title{Filtering markers by the relative aboundance of the most aboundant genotype}
\usage{
FilterMarkerByMajorGenotype(object, cutoff = NULL, print.n.markers = T)
}
\arguments{
\item{object}{\code{Genotype}. The object to be filtered.}

\item{cutoff}{\code{numeric}. Markers where relative number of calls having
the same genotype call is > \code{cutoff} will be masked.}

\item{print.n.markers}{logical. Should the number of remaining markers be
printed on the screen?}
}
\value{
a filtered \code{Genotype} object, with updated filter table and
\code{MarkerInfo$Filtered_at_step}
}
\description{
Filtering markers by the relative aboundance of the most aboundant genotype
}
\examples{
data("Mango_raw")
MarkerNo(MxS)
genotype_freq(AlleleCounting(MxS))["Contig4311_367", , drop=FALSE]
MxS = FilterMarkerByMajorGenotype(MxS, cutoff=0.95)
"Contig4311_367" \%in\% MarkerNames(MxS)
getMarkerInfo(MxS, Masked=F)["Contig4311_367", ]
}

