% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.R
\name{FilterMarkerByMinorGenotype}
\alias{FilterMarkerByMinorGenotype}
\title{Filtering markers by the aboundance of the least aboundant genotype}
\usage{
FilterMarkerByMinorGenotype(object, cutoff = NULL, print.n.markers = T)
}
\arguments{
\item{object}{\code{Genotype}. The object to be filtered.}

\item{cutoff}{\code{numeric}. Markers with a less than \code{cutoff}
absolute number of calls in some genotype will be masked.}

\item{print.n.markers}{logical. Should the number of remaining markers be
printed on the screen?}
}
\value{
a filtered \code{Genotype} object, with updated filter table and
\code{MarkerInfo$Filtered_at_step}
}
\description{
This function is used when you want to make sure all markers have the 3
possible genoytpe calls.
}
\examples{
data("Mango_raw")
MarkerNo(MxS)
genotype_freq(AlleleCounting(MxS))["Contig1327_496", , drop=FALSE]
MxS = FilterMarkerByMinorGenotype(MxS, cutoff=2)
"Contig1327_496" \%in\% MarkerNames(MxS)
getMarkerInfo(MxS, Masked=F)["Contig1327_496", ]
}

