% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.R
\name{FilterMarkerByNoCall}
\alias{FilterMarkerByNoCall}
\title{Filtering Markers by relative number of no calls}
\usage{
FilterMarkerByNoCall(object, cutoff = NULL, print.n.markers = T)
}
\arguments{
\item{object}{\code{Genotype}. The object to be filtered.}

\item{cutoff}{\code{numeric}. Markers with a relative number of missing calls
> cutoff will be masked.}

\item{print.n.markers}{logical. Should the number of remaining markers be
printed on the screen?}
}
\value{
a filtered \code{Genotype} object, with updated filter table and
\code{MarkerInfo$Filtered_at_step}
}
\description{
Filtering Markers by relative number of no calls
}
\examples{
data("Mango_raw")
MarkerNo(MxS)
MxS = FilterMarkerByNoCall(MxS, cutoff=0.1)
}

