% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.R
\name{FilterMarkerByPIC}
\alias{FilterMarkerByPIC}
\title{Filtering Markers by their polymorphic information content}
\usage{
FilterMarkerByPIC(object, cutoff = NULL, print.n.markers = T)
}
\arguments{
\item{object}{\code{Genotype}. The object to be filtered.}

\item{cutoff}{\code{numeric}. Markers with a PIC < cutoff will be masked.}

\item{print.n.markers}{logical. Should the number of remaining markers be
printed on the screen?}
}
\value{
a filtered \code{Genotype} object, with updated filter table and
\code{MarkerInfo$Filtered_at_step}
}
\description{
Filtering Markers by their polymorphic information content
}
\examples{
data("Mango_raw")
MarkerNo(MxS)
MxS = FilterMarkerByPIC(MxS, cutoff=0.1)
}

