% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.R
\name{FilterMarkerByQuality}
\alias{FilterMarkerByQuality}
\title{Quality filtering of markers}
\usage{
FilterMarkerByQuality(object, quality.criterion = "", cutoff = NULL,
  print.n.markers = T)
}
\arguments{
\item{object}{\code{Genotype}. The object to be filtered.}

\item{quality.criterion}{chatacter. The name of the quality criterion to
filter the markers by. One of the followings:
\describe{
  \item{no.calls}{filter markers with relative number of no calls > cutoff.}
  \item{pic}{filter markers with Polymorphic Information Content < cutoff.}
  \item{major.genotype}{filter markers with relative number of samples
  having the same call > cutoff. This is important to prevent a situation
  where almost all samples are heterozygous.}
  \item{minor.genotype}{filter markers where (at least) one of the 3
  possible genotype calls has relative number < cutoff. This is important
   when we want to make sure all 3 possible genotype calls exist for this
    marker.}
}}

\item{cutoff}{\code{numeric}. The cutoff for the filtration}

\item{print.n.markers}{logical. Should the number of remaining markers be
printed on the screen?}
}
\value{
The filtered \code{Genotype} object, with updated filter table and
marker info table.
}
\description{
A general function for several standard quality filtrations of markers
}
\examples{
data("Mango_raw")
MarkerNo(MxS)
MxS = FilterMarkerByQuality(MxS, quality.criterion="no.calls", cutoff=0.1)
MxS = FilterMarkerByQuality(MxS, quality.criterion="pic", cutoff=0.1)
Note: one marker is all heterozygous:
genotype_freq(AlleleCounting(MxS))["Contig4311_367", , drop=FALSE]
MxS = FilterMarkerByQuality(MxS, quality.criterion="major.genotype",
cutoff=0.9)
"Contig4311_367" \%in\% MarkerNames(MxS)
getMarkerInfo(MxS, Masked=F)["Contig4311_367", ]
genotype_freq(AlleleCounting(MxS))["Contig1309_880", , drop=FALSE]
MxS = FilterMarkerByQuality(MxS, quality.criterion="minor.genotype",
cutoff=1)
getMarkerInfo(MxS, Masked=F)["Contig1309_880", ]
}

