% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.R
\name{FilterSampleByInfo}
\alias{FilterSampleByInfo}
\title{Filtering samples by some \code{SampleInfo} column}
\usage{
FilterSampleByInfo(object, property.name = "", property.criterion = NULL,
  remove = "greaterOrEquall", print.n.samples = T)
}
\arguments{
\item{object}{\code{Genotype}. The object to be filtered.}

\item{property.name}{\code{character}. The name of the \code{SampleInfo}
column to use for filtering}

\item{property.criterion}{a single \code{character} or \code{numeric}
variable. If \code{SampleInfo$property.name} is a \code{character}, use a
\code{character} \code{property.criterion} to filter all samples with
\code{property.name} equal to \code{property.criterion} or filter all
samples with \code{property.name} different from \code{property.criterion},
depending on \code{remove}. If \code{SampleInfo$property.name} is a
\code{numeric} vector, use a \code{numeric} \code{property.criterion} as a
cutoff.}

\item{remove}{\code{character}. One of the following:
    "equal", "different", "greater", "greaterOrEquall", "smaller" or
"smallerOrEquall".
Which samples to remove with respect to \code{property.criterion}}

\item{print.n.samples}{logical. Should the number of remaining samples be
printed on the screen?}
}
\value{
a filtered \code{Genotype} object, with updated filter table and
\code{SampleInfo$Filtered_at_step}
}
\description{
This functions allows you to filter samples by any property. To use it, the
 relevant property should appear in the object's \code{SampleInfo} table.
}
\examples{
Leaving only samples of graft strains
data("Mango_raw")
SampleNo(MxS)
MxS = FilterSampleByInfo(MxS, property.name="Rootstock/Graft",
property.criterion="G", remove="different")

}

