% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.R
\name{FilterSampleByNoCall}
\alias{FilterSampleByNoCall}
\title{Filtering samples by relative number of no calls}
\usage{
FilterSampleByNoCall(object, cutoff = NULL, print.n.samples = T)
}
\arguments{
\item{object}{\code{Genotype}. The object to be filtered}

\item{cutoff}{\code{numeric}. Samples with relative number of missing calls
higher than this value will be filtered.}

\item{print.n.markers}{logical. Should the number of remaining markers be
printed on the screen?}
}
\value{
The same object with the samples with too many missing calls
filtered.
}
\description{
Filtering samples by relative number of no calls
}
\examples{
data("Mango_raw")
MarkerNo(MxS)
MxS = FilterSampleByNoCall(MxS, cutoff=0.3)
}

