% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.R
\name{FilterSampleByQuality}
\alias{FilterSampleByQuality}
\title{Quality filtering of samples}
\usage{
FilterSampleByQuality(object, quality.criterion = "", cutoff = NULL,
  print.n.samples = T)
}
\arguments{
\item{object}{\code{Genotype}. The object to be filtered.}

\item{quality.criterion}{chatacter. The name of the quality criterion to
filter the samples by. Currently there is only one acceptable value:
   "no.calls" - filter samples with relative number of no calls > cutoff.}

\item{cutoff}{\code{numeric}. The cutoff for the filtration}

\item{print.n.samples}{logical. Should the number of remaining samples be
printed on the screen?}
}
\value{
The filtered \code{Genotype} object, with updated filter table and
sample info table.
}
\description{
A general function for several standard quality filtrations of samples
}
\examples{
data("Mango_raw")
SampleNo(MxS)
MxS = FilterSampleByQuality(MxS, quality.criterion="no.calls", cutoff=0.3)
}

