% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basics.R
\docType{class}
\name{filterTable-class}
\alias{filterTable-class}
\title{A data frame for documenting the filtering process}
\description{
A data frame with the name of each filtering step, and the numbers of
markers and samples following that step. The first row in the table shows
the numbers of markers and samples in the raw, unfiltered data.
}
\section{Slots}{

\describe{
\item{\code{filter_table}}{data.frame: The table of filtration information}
}}

