% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Identical_samples.R
\name{FindIdentical}
\alias{FindIdentical}
\title{Find pairs of identical samples}
\usage{
FindIdentical(object, error.freq = 0)
}
\arguments{
\item{object}{\code{Genotype}. The name of the object to find the identical
samples in.}

\item{error.freq}{\code{numeric}. A single numeric value: the estimated
genotype call error frequency.}
}
\value{
\code{IdenticalSamples}. A table of identical pairs.
}
\description{
finding pairs of individuals which are genetically identical accross the
panel of marker set
}
\details{
\code{FindIdentical} considers any pair of samples, that have
identical genotype call in \code{1-error.freq} or more markers, as
genetically identical.
}
\examples{
data("Mango_row")
F0 = FindIdentical(MxS)
F0.05 = FindIdentical(MxS, error.freq=0.05)
}
\seealso{
\code{\link{IdenticalGroups}} for creating groups of identical
samples, \code{\link{AddGroup2SampleInfo}} for adding the identity group to
the sample information, and \code{\link{Unique}} for filtering all samples
but one from each identity group.
}

