% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{FindParentOffspring}
\alias{FindParentOffspring}
\title{Find potential parent-offspring pairs}
\usage{
FindParentOffspring(object, error.freq = 0, MaskedSample = T)
}
\arguments{
\item{object}{\code{Genotype}. The object to find selfings in.}

\item{error.freq}{\code{numeric}. A value between 0 and 1, with the
estimated genotype call error rate.}
}
\value{
\code{data.frame}. A table with each row representing a potential
pair of parent and offpring. The columns are:
\describe{
  \item{parent}{The potential parent}
  \item{offspring}{The potential offspring}
  \item{agreement}{The relative number of markers in agreement with this
   possibility}
  \item{NMarkers}{The total number of markers compared}
  \item{identity}{The relative number of markers identical in both samples}
  \item{expected_identity}{The expected relative number of markers identical
  between the parent and a potential offspring.}
}
}
\description{
Finding potential pairs of a parent and an offspring, which is the result of
the parent selfing, within a \code{Genotype} object.
}
\examples{
data("citrus_clean")
PO = FindParentOffspring(MxS, error.freq=0.05)
}

