% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{FindSelfing}
\alias{FindSelfing}
\title{Find potential selfings}
\usage{
FindSelfing(object, error.freq = 0)
}
\arguments{
\item{object}{\code{Genotype}. The object to find selfings in.}

\item{error.freq}{\code{numeric}. A value between 0 and 1, with the
estimated genotype call error rate.}
}
\value{
\code{data.frame}. A table with potential selfings. Each row
represents a pair of selfing parent and its offpring. The columns are:
\describe{
  \item{selfing_parent}{The potential selfing parent}
  \item{offspring}{The potential offspring}
  \item{agreement}{The relative number of markers in agreement with this
   possibility}
  \item{NMarkers}{The total number of markers compared}
  \item{identity}{The relative number of markers identical in both samples}
  \item{parent_heterozygousity}{The relative number of heterozygous loci in
   the potential parent}
  \item{offspring_heterozygosity}{The relative number of heterozygous loci
  in the potential offspring}
  \item{generation}{The total number selfing generations, as calculated from
   the reduction in heterozygousity.}
}
}
\description{
Finding potential pairs of a parent and an offspring, which is the result of
the parent selfing, within a \code{Genotype} object.
}
\examples{
data("citrus_clean")
self = FindSelfing(MxS, error.freq=0.05)
}

