% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{Fis}
\alias{Fis}
\alias{Fit}
\alias{Fst}
\title{Fixation index: F-statistics}
\usage{
Fis(object)

Fst(object)

Fit(object)
}
\arguments{
\item{object.}{\code{Populations}. The object to calculate the
inbreeding coefficient in.}
}
\value{
\code{numeric}. A numeric vector, with the inbreeding coefficient
in each marker.
}
\description{
The inbreeding coefficient, as the reduction in heterozgous frequency
compared to the expected frequency, due to inbreeding within each sub-
population (Fis), due to the effect of subpopulations (Fst), and globaly
among the whole population (Fit)
}
\examples{
in random sub-populations:
data("citrus_clean")
group.size= round(SampleNo(MxS)/3)
Total = 1:SampleNo(MxS)
group.samples=list()
for (i in 1:3) {
  if (i<3) {
    group.samples[[i]]=sample(Total,group.size)
    Total = Total[!Total\%in\%group.samples[[i]]]
  } else group.samples[[i]]= Total
}
group.samples = lapply(group.samples, function(x) SampleNames(MxS)[x])
A1 = AlleleCounting(SetSampleFilter(MxS, unlist(group.samples[-1])))
A2 = AlleleCounting(SetSampleFilter(MxS, unlist(group.samples[-2])))
A3 = AlleleCounting(SetSampleFilter(MxS, unlist(group.samples[-3])))
RandomPop = new("Populations",Pop=list(A1, A2, A3))
fis = Fis(RandomPop)
fst = Fst(RandomPop)
fit = Fit(RandomPop)
summary(fis)
summary(fst)
summary(fit)

in real sub-populations:
data("Avocado_populations")
fis = Fis( )
fst = Fst(AvPop)
fit = Fit(AvPop)
summary(fis)
summary(fst)
summary(fit)
}

