% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{Fs}
\alias{Fs}
\title{Fixation index}
\usage{
Fs(object)
}
\arguments{
\item{object.}{\code{AlleleStat}. The object to calculate the fixation index
from.}
}
\value{
\code{numeric}. A numeric vector, with the fixation index for each
marker.
}
\description{
Fixation index, as the reduction in heterozygousity, compared to expectd
heterozygousity in a random mating population
}
\examples{
data("citrus_clean")
MarkersStat = AlleleCounting(MxS)
fs = Fs(MarkersStat)
}
\seealso{
\code{\link{pX}} and \code{\link{pY}} for allele frequencies,
\code{\link{Hobs}} for observed heterozygousity and  \code{\link{Hexp}} for
observed heterozygousity.
}

