% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust_tree.R
\name{GeneticDistance}
\alias{GeneticDistance}
\title{Genetic distance between samples}
\usage{
GeneticDistance(object, method = c("euclidean", "PSA", "IBD"), maf = 0.05,
  missing.rate = 0.05)
}
\arguments{
\item{object}{\code{Genotype}. The object to calculate the distance in.}

\item{method}{\code{character}. The method of distance. One of the
following:
#' \describe{
  \item{"euclidean"}{euclidean distance between 2 vectors of genotype calls,
   where XX is translated to 1, XY to 2, and YY to 3.}
  \item{"PSA"}{proportion of shared alleles, and}
  \item{"IBD"}{kinship, where kinship is calculated using \code{SNPRelate}
  package \code{snpgdsIBDKING} function.}
}}

\item{maf}{\code{numeric}. Minor allele frequency threshold for
\code{snpgdsIBDKING}.}

\item{missing.rate}{\code{numeric}. Use SNPs with "<= missing.rate" only in
\code{snpgdsIBDKING}.}
}
\value{
\code{dist}
}
\description{
Genetic distance between samples
}
\details{
This function is used by \code{\link{bootstrap}}, which uses
genetic distance to build an hierarchical clustered tree.
}
\examples{
data("citrus_clean")
De = GeneticDistance(MxS, method="euclidean")
Dp = GeneticDistance(MxS, method="PSA")
Di = GeneticDistance(MxS, method="IBD")
}

