% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basics.R
\docType{class}
\name{Genotype-class}
\alias{Genotype-class}
\title{A S4 class for a Genotype database}
\description{
A class that typically represents a genotyping assay. It contains (by
inharitance) a GenotypeCalls object, which is the genotype call matrix, a
SampleInfo and a MarkerInfo objects, which hold the information about the
samples and the markers in the assay, and the filterTable, which documents
the cleaning of the assay data.
}
\section{Slots}{

\describe{
\item{\code{MarkerFilter}}{integer: The indeces of markers that were not masked by
the filtration process.}

\item{\code{SampleFilter}}{integer: The indeces of samples that were not masked by
the filtration process.}
}}

