% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{genotype_freq}
\alias{genotype_freq}
\title{Genotype frequencies}
\usage{
genotype_freq(object)
}
\arguments{
\item{object}{\code{AlleleStat}. The object to get the genotype frequencies
from.}
}
\value{
\code{matrix}. The matrix of genotype frequencies. Each row
corresponds to a marker, and the columns list the number of individual
samples with "XX", "XY" and "YY" genotypes.
}
\description{
Genotype frequencies
}
\examples{
data("Mango_raw")
MarkersStat = AlleleCounting(MxS)
GF = genotype_freq(MarkersStat)
head(GF)
}

