% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basics.R
\docType{class}
\name{GenotypeCalls-class}
\alias{GenotypeCalls-class}
\title{A S4 class for a matrix of genotype calls.}
\description{
A matrix of genotype calls, with each row representing a marker, and each
column representing an individual sample. The genotype call is in the form
of a character.
}
\section{Slots}{

\describe{
\item{\code{Calls}}{matrix: The matrix of genotype calls}

\item{\code{Samples}}{character: Individual samples IDs of the}

\item{\code{Markers}}{character: Markers IDs}
}}

