% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STRUCTURE_interface.R
\name{get_matrix}
\alias{get_matrix}
\title{Getting the matrix of sub-population ancestry from STRUCTURE output}
\usage{
get_matrix(f, path, k, print_file)
}
\arguments{
\item{f}{\code{character}. The file name}

\item{path}{\code{character}. The path to the directory where the file is.}

\item{k}{\code{numeric}. The number of sub-population or 'clusters'. This
number is a pre-defined parameter required by STRUCTURE.}

\item{print_file}{\code{logical}. Should the name of the file be printed?}
}
\value{
the matrix of relative ancestry.
}
\description{
STRUCTURE calculates for each individual sample, how much of its genome
belongs to each one of K sub-populations, or so-called clusters. This
function extracts this information from the STRUCTURE output file, and
returns a matrix with K columns and n rows, n being the number of samples.
   This function is used by \code{\link{Sum_Structure_Results}}, which sums
over multiple simulations with the same data and K.
}

