% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basics.R
\name{getCalls}
\alias{getCalls}
\title{Get a matrix of genotype calls}
\usage{
getCalls(object, MaskedMarker = TRUE, MaskedSample = TRUE)
}
\arguments{
\item{MaskedMarker}{logical. Whether to seclude masked markers}

\item{MaskedSample}{logical. Whether to seclude masked samples}

\item{object.}{The object to get the genotype calls from}
}
\value{
\code{matrix} of genotype calls:
\describe{
\item{XX}{homozygous to the 1st allele}
\item{XY}{heterozygous}
\item{YY}{homozygous to the 2nd allele}
\item{NA}{genotype call not available}
}
}
\description{
Get a matrix of genotype calls
}
\examples{
data("Mango_raw")
a = getCalls(MxS)
SampleNames(MxS, Masked=F)
}

