% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_table.R
\name{getFilterTable}
\alias{getFilterTable}
\title{Getting the filter table}
\usage{
getFilterTable(object)
}
\arguments{
\item{object}{\code{Genotype} or \code{filterTable}. The object to get the
filter table from}
}
\value{
The filter table
}
\description{
Getting the filter table
}

