% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.R
\name{GetMarkerFilter}
\alias{GetMarkerFilter}
\title{Getting the marker Filter}
\usage{
GetMarkerFilter(object)
}
\arguments{
\item{object}{\code{Genotype}. The object to get the filter from.}
}
\value{
The indeces of the unmasked markers.
}
\description{
Getting the marker Filter
}

