% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basics.R
\name{getMarkerInfo}
\alias{getMarkerInfo}
\title{Marker information}
\usage{
getMarkerInfo(object, Masked = TRUE)
}
\arguments{
\item{Masked}{logical. Whether to avoid getting the information from masked
markers too}

\item{object.}{The object to get the markers information from.}
}
\value{
\code{MarkerInfo}
}
\description{
Marker information
}
\examples{
data("Mango_raw")
MI = getMarkerInfo(MxS)
head(MI)
}

