% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.R
\name{GetSampleFilter}
\alias{GetSampleFilter}
\title{Getting the sample Filter}
\usage{
GetSampleFilter(object)
}
\arguments{
\item{object}{\code{Genotype}. The object to get the filter from.}
}
\value{
The indeces of the unmasked samples.
}
\description{
Getting the sample Filter
}

