% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basics.R
\name{getSampleInfo}
\alias{getSampleInfo}
\title{Sample information}
\usage{
getSampleInfo(object, Masked = TRUE)
}
\arguments{
\item{Masked}{logical. Whether to avoid getting the information from masked
samples}

\item{object.}{The object to get the samples information from.}
}
\value{
\code{SampleInfo}
}
\description{
Sample information
}
\examples{
data("Mango_raw")
SI = getSampleInfo(MxS)
head(SI)
}

