% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust_tree.R
\name{getTreeLeafOrder}
\alias{getTreeLeafOrder}
\title{Getting the order of leaves in a 'phylo' tree}
\usage{
getTreeLeafOrder(object)
}
\arguments{
\item{object}{\code{phylo}. The tree of interest}
}
\value{
\code{integer}. The order of the leaves in the tree.
}
\description{
This allows you to re-order a Genotype object according to the similarity
between samples
}
\examples{
data("Avocado_clean123")
library(ape)
data("Avocado_phylo")
o = getTreeLeafOrder(b.tree)
MxS = SetSampleOrder(MxS, Order=o)
}

