% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{Hexp}
\alias{Hexp}
\title{The expected frequency of heterozygous}
\usage{
Hexp(object)
}
\arguments{
\item{object.}{\code{AlleleStat}. The object to calculate heterozygousity
frequency from.}
}
\value{
\code{numeric}. A numeric vector, with the expected frequency of
heterozygous for each marker.
}
\description{
Calculates the expected relative frequency of heterozygous, per each marker.
}
\examples{
data("citrus_clean")
MarkersStat = AlleleCounting(MxS)
h = Hexp(MarkersStat)
}
\seealso{
\code{\link{pX}} and \code{\link{pY}} for allele frequencies,
\code{\link{Hobs}} for observed heterozygousity.
}

