% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{Hi}
\alias{Hi}
\alias{Hs}
\alias{Ht}
\title{Heterozygosity in a global population containing several sub-populations}
\usage{
Hi(object)

Hs(object)

Ht(object)
}
\arguments{
\item{object.}{\code{Populations}. The object to calculate the
heterozygosity in.}
}
\value{
\code{numeric}. A numeric vector, with the heterozygosity in each
marker.
}
\description{
Global observed (\code{Hi}) heterozygousity, global expected (\code{Ht})
heterozygousity and weighted mean expected heterozygousity among
sub-populations (\code{Hs})
}
\examples{
data("citrus_clean")
group.size= round(SampleNo(MxS)/3)
Total = 1:SampleNo(MxS)
group.samples=list()
for (i in 1:3) {
  if (i<3) {
    group.samples[[i]]=sample(Total,group.size)
    Total = Total[!Total\%in\%group.samples[[i]]]
  } else group.samples[[i]]= Total
}
group.samples = lapply(group.samples, function(x) SampleNames(MxS)[x])
A1 = AlleleCounting(SetSampleFilter(MxS, unlist(group.samples[-1])))
A2 = AlleleCounting(SetSampleFilter(MxS, unlist(group.samples[-2])))
A3 = AlleleCounting(SetSampleFilter(MxS, unlist(group.samples[-3])))
RandomPop = new("Populations",Pop=list(A1, A2, A3))
hi = Hi(RandomPop)
ht = Ht(RandomPop)
hs = Hs(RandomPop)
summary(abs(hi - Hobs(AlleleCounting(MxS))))
summary(abs(ht - Hexp(AlleleCounting(MxS))))
summary(abs(hs - Hexp(AlleleCounting(MxS))))
}

