% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{Hobs}
\alias{Hobs}
\title{The observed frequency of heterozygous}
\usage{
Hobs(object)
}
\arguments{
\item{object.}{\code{AlleleStat}. The object to calculate heterozygousity
frequency from.}
}
\value{
\code{numeric}. A numeric vector, with the frequency of the
heterozygous for each marker.
}
\description{
Calculates the relative frequency of heterozygous, per each marker.
}
\examples{
data("citrus_clean")
MarkersStat = AlleleCounting(MxS)
h = Hobs(MarkersStat)
}
\seealso{
\code{\link{pX}} and \code{\link{pY}} for allele frequencies,
\code{\link{Hexp}} for expected heterozygousity.
}

