% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Identical_samples.R
\name{IdenticalGroups}
\alias{IdenticalGroups}
\title{Grouping together identical samples}
\usage{
IdenticalGroups(id_table)
}
\arguments{
\item{id_table}{\code{IdenticalSamples}. A special kind of data.frame which
specifies pairs of identical samples}
}
\value{
\code{list}. A list of \code{character} vectors. Each one is a set
of sample names which are all genetically identical
}
\description{
Grouping together identical samples
}
\details{
\code{IdenticalGroups} uses the \code{MCL} package \code{mcl}
function to partition the samples into groups.
}
\examples{
data("Mango_row")
F0.05 = FindIdentical(MxS, error.freq=0.05)
group = IdenticalGroups(F0.05)
}
\seealso{
\code{\link{FindIdentical}} for finding pairs of identical samples,
, \code{\link{AddGroup2SampleInfo}} for adding the identity group to the
sample information and \code{\link{Unique}} for filtering all samples but
one from each identity group.
}

