% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Identical_samples.R
\docType{class}
\name{IdenticalSamples-class}
\alias{IdenticalSamples-class}
\title{A \code{data.frame} for a list of pairs of genetically identical samples}
\description{
A data frame with the following variables:
   sample1: the first sample in the identical pair
   sample2: the second sample in the identical pair
   identity: the relative number of markers in which the pair is identical.
}
\seealso{
\code{\link{FindIdentical}} for finding pairs of identical samples,
\code{\link{IdenticalGroups}} for identifying groups of identical samples,
 \code{\link{AddGroup2SampleInfo}} for adding the identity group to the
 sample information, and \code{\link{Unique}} for filtering all samples but
 one from each identity group.
}

