% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hclust_tree.R
\name{invert_tree}
\alias{invert_tree}
\title{Inverting the order of a 'phylo' tree or a branch of it}
\usage{
invert_tree(tree1, node = NULL)
}
\arguments{
\item{tree1}{\code{phylo}. The tree to invert}

\item{node}{\code{numeric}. The node number around which to invert the tree.
If this argument is missing or NULL, the entire tree will be inverted.}
}
\value{
\code{phylo}. The re-ordered tree.
}
\description{
This function allows you to invert the order of leaves in a 'phylo' object.
You can invert the whole tree, or just all the leaves that descend from a
certain node. It is sometimes desireable when plotting the tree.
}
\examples{
library("ape")
data("Avocado_phylo")
c.tree = invert_tree(b.tree)
d.tree = invert_tree(b.tree, 128)
plot(b.tree)
plot(c.tree)
plot(d.tree)
}
\seealso{
\code{\link{bootstrap}}, \code{\link{GeneticDistance}},
\code{\link{getTreeLeafOrder}}, \code{\link{DrowPopulationTree}}
}

