% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBD.R
\name{kinship}
\alias{kinship}
\title{Calculate kinship using \code{SNPRelate} package}
\usage{
kinship(object, method = c("King", "MoM"), maf = 0.05,
  missing.rate = 0.05, rm.negatives = T)
}
\arguments{
\item{object}{\code{Genotype}. The object to calculate kinship from}

\item{method}{\code{character}. The method to use for identity-by-descent
calculation. Either "King" for using the "KING-robust" method, or "MoM" for
PLINK method of moment.}

\item{maf}{\code{numeric}. Cutoff for minor allele frequency.}

\item{missing.rate}{\code{numeric}. Cutoff for missing call rate.}

\item{rm.negatives}{\code{logical}. Should negative values be changed to 0?}
}
\value{
\code{numeric}. the estimated kinship coefficients.
}
\description{
Calculate kinship using \code{SNPRelate} package
}

