% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cervus_interface.R
\name{Marker_Match_from_Cervus}
\alias{Marker_Match_from_Cervus}
\title{Using known parent-offspring pairs to filter non-matching markers}
\usage{
Marker_Match_from_Cervus(object, genotype, mm_cutoff = Inf,
  lod_cutoff = -Inf)
}
\arguments{
\item{object}{\code{data.frame}. A table with Cervus results. Use
\code{\link{read.cervus}} to obtain it.}

\item{genotype}{\code{Genotype} or \code{matrix}. The object to take the
genotype calls from.}

\item{mm_cutoff}{\code{numeric}. The maximum number of mismatches between an
offspring and a parent, to count them as a true pair.}

\item{lod_cutoff}{\code{numeric}. The minumum lod score of an offspring-
parent pair, to count them as a true pair.}
}
\value{
\code{numeric}. A vector the same length as the number of markers,
with the relative number of true parent-offspring pair that match according
to each marker.
}
\description{
If we have previous knowledge of parent-offspring pairs in our data (usually
mother-offspring) we can use this knowledge to filter out markers that don't
support it. A parent offspring pair must have a common allele. If a marker
does not match in a significant number of known and well established parent-
offspring pair, this marker may not be trusted. Provided that you have a
large enough number of previously known parent-offspring pairs, you can run
'Cervus' software with known parent on your genotype data, and then use this
 function to get the match of each marker to these pairs.
}
\examples{
data("known_mother")
data("citrus_clean")
marker_match = Marker_Match_from_Cervus(known_mother, genotype=MxS, mm_cutoff=5,
lod_cutoff=10)
}
\seealso{
\code{\link{write.cervus}}, \code{\link{write.candidates}},
\code{\link{read.cervus}}
}

