% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basics.R
\docType{class}
\name{MarkerInfo-class}
\alias{MarkerInfo-class}
\title{A S4 class for marker information}
\description{
A data frame which row names are the markers IDs. It can include any marker
information available, such as the two alleles, genome positions if
available, and the quality ranks given by the Fluidigm team.
}
\section{Slots}{

\describe{
\item{\code{MInfo}}{data.frame: The table of marker info}

\item{\code{Markers}}{character: Marker IDs. identical to the row names of MInfo.}
}}

