% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basics.R
\name{MarkerNames}
\alias{MarkerNames}
\title{The marker names}
\usage{
MarkerNames(object, Masked = TRUE)
}
\arguments{
\item{Masked}{logical. Whether to seclude masked markers}

\item{object.}{The object to get the markers from}
}
\value{
\code{character} the marker names
}
\description{
The marker names
}
\examples{
data("Mango_raw")
MarkerNames(MxS)
MarkerNames(MxS, Masked=F)
}

