% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basics.R
\name{MarkerNo}
\alias{MarkerNo}
\title{The number of markers}
\usage{
MarkerNo(object, Masked = TRUE)
}
\arguments{
\item{Masked}{logical. Whether to ignore masked markers}

\item{object.}{The object to count the markers from}
}
\value{
\code{numeric} the number of markers
}
\description{
The number of markers
}
\examples{
data("Mango_raw")
MarkerNo(MxS)
MarkerNo(MxS, Masked=F)
}

