% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{MinimalSet}
\alias{MinimalSet}
\title{Find a minimal set of markers to distinguish between all samples}
\usage{
MinimalSet(object, marker.score = NULL, filter.nocall = T)
}
\arguments{
\item{object}{\code{Genotype}. The object conatining the genotype calls to
use.}

\item{marker.score}{\code{numeric}. Optional. A vector of marker score. In
case there is a choice between several markers, a marker with a higher score
will be preferred.}

\item{filter.nocall}{\code{character}. Should markers with missing calls be
filtered? options are:
\describe{
\item{quantile}{filter out the 25\% of markers with highest rate of no call}
\item{all}{filter out the markers with any missing call}
\item{none}{no filtering.}
}}
}
\value{
\code{character}. The names of the markers in the minimal set.
}
\description{
Find a minimal set of markers to distinguish between all samples
}
\examples{
data("citrus_clean")
myset = MinimalSet(MxS)

}

