% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{ocPol}
\alias{ocPol}
\title{Intercross pollination}
\usage{
ocPol(object, test = character(), N = numeric())
}
\arguments{
\item{object}{\code{AlleleStat}. The object to calculate the polymorphic
information content from.}

\item{test}{\code{character}. a vector of genotype calls for the same
markers that exist in \code{object}.}

\item{N}{\code{numeric}. A single numeric value. The maximum number of
markers to test}
}
\value{
\code{data.frame}. A table with maximum N rows corresponding to
markers, in which test is homozygous for the less frequent allele in the
object.
}
\description{
Intercross pollination
}
\examples{
data("citrus_clean")
MarkersStat = AlleleCounting(MxS)
test = getCalls(MxS)[, 1]
ocpol = ocPol(MarkersStat, test=test, N=10)
}

