% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{pair_common_alleles}
\alias{pair_common_alleles}
\title{Markers with shared alleles between a pair of samples}
\usage{
pair_common_alleles(v1, v2)
}
\arguments{
\item{v1}{\code{character}. a vector of genotype calls for one sample.}

\item{v2}{\code{character}. a vector of genotype calls for the other sample.}
}
\value{
\code{logical}. A vector of the same length as the number of markers
with no missing calls in both samples, specifying for each marker whether
the two samples have a common allele.
}
\description{
Markers with shared alleles between a pair of samples
}

