% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{PIC}
\alias{PIC}
\title{Polymorphic information content}
\usage{
PIC(object, type = 1)
}
\arguments{
\item{object.}{\code{AlleleStat}. The object to calculate the polymorphic
information content from.}
}
\value{
\code{numeric}. A numeric vector, with the polymorphic information
content for each marker.
}
\description{
Polymorphic information content
}
\examples{
data("citrus_clean")
MarkersStat = AlleleCounting(MxS)
pic = pic(MarkersStat)
}

