% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{PopNo}
\alias{PopNames}
\alias{PopNo}
\title{Number of populatoins}
\usage{
PopNo(object)

PopNames(object)
}
\arguments{
\item{object.}{\code{Populations}. The object with the populations.}
}
\description{
The number or names of populations in a \code{Populations} object
}

