% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{PopSize}
\alias{PopSize}
\title{Number of individuals with a genotype call}
\usage{
PopSize(object)
}
\arguments{
\item{object.}{\code{AlleleStat}. The object to count the samples in.}
}
\value{
\code{numeric}. A numeric vector, with the number of genotype calls
available for each marker.
}
\description{
Counting the number of samples with valid genotype calls, per each marker.
}
\examples{
data("citrus_clean")
MarkersStat = AlleleCounting(MxS)
pz = PopSize(MarkersStat)
SampleNo(MxS)
summary(pz)
}

