% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\docType{class}
\name{Populations-class}
\alias{Populations-class}
\title{A S4 class for a sub-populations.}
\description{
A S4 class for a sub-populations.
}
\section{Slots}{

\describe{
\item{\code{Pop}}{\code{list}: A named list of \code{AlleleStat} objects, each
specifying allele genotype frequencies of the same markers in a different
population.}
}}

