% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{pX}
\alias{pX}
\title{Allele 1 frequency}
\usage{
pX(object)
}
\arguments{
\item{object.}{\code{AlleleStat}. The object to calculate allele frequency
from.}
}
\value{
\code{numeric}. A numeric vector, with the frequency of the 1st
allele for each marker.
}
\description{
Calculates the frequency of the 1st, per each marker.
}
\examples{
data("citrus_clean")
MarkersStat = AlleleCounting(MxS)
px = pX(MarkersStat)
}
\seealso{
\code{\link{pY}}
}

