% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{pY}
\alias{pY}
\title{Allele 2 frequency}
\usage{
pY(object)
}
\arguments{
\item{object.}{\code{AlleleStat}. The object to calculate allele frequency
from.}
}
\value{
\code{numeric}. A numeric vector, with the frequency of the 2nd
allele for each marker.
}
\description{
Calculates the frequency of the 2nd, per each marker.
}
\examples{
data("citrus_clean")
MarkersStat = AlleleCounting(MxS)
py = pY(MarkersStat)
}
\seealso{
\code{\link{pX}}
}

