% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cervus_interface.R
\name{read.cervus}
\alias{read.cervus}
\title{Read 'Cervus' results}
\usage{
read.cervus(file_path)
}
\arguments{
\item{file_path}{\code{character}. The path and file name of the citrus .csv
file.}
}
\value{
\code{data.frame}. A table with the cervus results.
}
\description{
Cervus is a computer program for assignment of parents to their offspring
using genetic markers. It first calculates allele frequencies from a
genotype call file. Than it runs a simulation to determine the level of log
likelihood and delta required for parental assignment. Finally it assigns
uses a list of offsprings, and a list of candidate parents to each one of
them, to assign parents to offsprings, both when one of the parents is
known, and without a known parent.
Cervus outputs a .csv file where every row specifies one offspring-mother-
father trio, and includes the number of mismatching markers, log likelihood
and delta
}
\seealso{
\code{\link{write.cervus}}, \code{\link{write.candidates}}
}

