% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{read.genotype}
\alias{read.genotype}
\title{Import a table of genotype calls.}
\usage{
read.genotype(file, object)
}
\arguments{
\item{file}{character: the name of the file to read the data from.}
}
\value{
a \code{Genotype} object
}
\description{
Import a table of genotype calls and create a \code{Genotype} object.
}
\examples{
MxS = read.genotype("Fluidigm_Output.csv")
}

