% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{read.markers}
\alias{read.markers}
\title{Import the marker data.}
\usage{
read.markers(object, file)
}
\arguments{
\item{object}{the object (typically of class \code{Genotype}) to add the
marker data into.}

\item{file}{character: the name of the file to read the marker data from.
The 1st column in the file should contain the marker IDs.}
}
\value{
the object containing the marker data.
}
\description{
Import a table of marker data and create a \code{MarkerInfo} object.
}

