% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{read.samples}
\alias{read.samples}
\title{Import the sample data.}
\usage{
read.samples(object, file)
}
\arguments{
\item{object}{the object (typically of class \code{Genotype}) to add the
sample data into.}

\item{file}{character: the name of the file to read the sample data from.
The 1st column in the file should contain the sample IDs.}
}
\value{
the object containing the sample data.
}
\description{
Import a table of sample data and create a \code{SampleInfo} object.
}

