% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.R
\name{RemoveFilter}
\alias{RemoveFilter}
\title{Remove a specific filter from a a \code{Genotype} object.}
\usage{
RemoveFilter(object, step_name = "no_step", print.n = T)
}
\arguments{
\item{object}{\code{Genotype}. The object to remove the filter from.}

\item{step_name}{character. The name of the filtration step to be cancelled.}

\item{print.n.markers}{logical. Should the new number of remaining markers
be printed on the screen?}
}
\value{
The \code{Genotype} object with the newly unfiltered markers or
samples.
}
\description{
All samples or markers that were removed by this filtration step will be
un-masked. Use this carefully. If other filtration steps were taking place
after the step that is being removed, the un-masked markers or samples will
not be affected by them.
}
\examples{
data("citrus_clean")
MarkerNo(MxS)
MxS = RemoveFilter(MxS, "Filtering out SNPs with a Category >=3")
getFilterTable(MxS)
}

