% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.R
\name{ResetMarkerFilter}
\alias{ResetMarkerFilter}
\title{UnFiltering the markers}
\usage{
ResetMarkerFilter(object)
}
\arguments{
\item{object}{\code{Genotype}. The object to be un-filtered.}
}
\value{
The unfiltered \code{Genotype} object
}
\description{
\code{ResetMarkerFilter} reset marker filter, so that all markers are
 unmasked
}
\examples{
data("Mango_raw")
MarkerNo(MxS)
MxS = SetMarkerFilter(MxS, Filter=c("Contig1987_1443", "Contig2006_1970",
    "mango_c15408_840"), step_name="arbitrary filtering")
getMarkerInfo(MxS, Masked=F)["Contig1987_1443", ]
MxS = ResetMarkerFilter(MxS)
}

