% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.R
\name{ResetSampleFilter}
\alias{ResetSampleFilter}
\title{UnFiltering the samples}
\usage{
ResetSampleFilter(object)
}
\arguments{
\item{object}{\code{Genotype}. The object to be un-filtered.}
}
\value{
The unfiltered (with respect to samples) \code{Genotype} object.
}
\description{
\code{ResetSampleFilter} reset sample filter, so that all samples are
 unmasked
}
\examples{
data("Mango_raw")
SampleNo(MxS)
MxS = SetSampleFilter(MxS, Filter=c("Contig1987_1443", "Contig2006_1970",
    "mango_c15408_840"), step_name="arbitrary filtering")
getSampleInfo(MxS, Masked=F)["Contig1987_1443", ]
MxS = ResetSampleFilter(MxS)
}

