% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basics.R
\docType{class}
\name{SampleInfo-class}
\alias{SampleInfo-class}
\title{A S4 class for sample information}
\description{
A data frame which row names are the sample IDs. It can include any sample
information available. Usually it will have a column with the cultivar name.
}
\section{Slots}{

\describe{
\item{\code{SInfo}}{data.frame: The table of sample info}

\item{\code{Samples}}{character: Sample IDs. identical to the row names of SInfo.}
}}

