% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basics.R
\name{SampleNames}
\alias{SampleNames}
\title{The sample names}
\usage{
SampleNames(object, Masked = TRUE)
}
\arguments{
\item{Masked}{logical. Whether to seclude masked samples}

\item{object.}{The object to get the samples from}
}
\value{
\code{character} the sample names
}
\description{
The sample names
}
\examples{
data("Mango_raw")
SampleNames(MxS)
SampleNames(MxS, Masked=F)
}

