% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basics.R
\name{SampleNo}
\alias{SampleNo}
\title{The number of samples}
\usage{
SampleNo(object, Masked = TRUE)
}
\arguments{
\item{Masked}{logical. Whether to ignore masked samples}

\item{object.}{The object to count the samples from}
}
\value{
\code{numeric} the number of samples
}
\description{
The number of samples
}
\examples{
data("Mango_raw")
SampleNo(MxS)
SampleNo(MxS, Masked=F)
}

