% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basics.R
\name{SetMarkerName}
\alias{SetMarkerName}
\title{Change the marker names}
\usage{
SetMarkerName(object, marker.name = character(), Masked = F)
}
\arguments{
\item{object}{\code{Genotype}. The object to change the markers names in.}

\item{marker.name}{code{character}. A character vector with the new marker
names.}
}
\value{
The object with the markers having their new names.
}
\description{
Change the marker names
}

