% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basics.R
\name{SetMarkerOrder}
\alias{SetMarkerOrder}
\title{Reorder the markers}
\usage{
SetMarkerOrder(object, Order = NULL)
}
\arguments{
\item{object}{\code{Genotype}. The object to reorder the markers in}

\item{Order}{\code{numeric} or \code{character}. A numeric vector with the
order of markers, or a character vector with the marker names ordered as
desired.}
}
\value{
The object with the markers re-ordered as specified.
}
\description{
Reorder the markers
}
\examples{
data("Mango_raw")
a = MarkerNames(MxS)
b = MarkerNames(SetMarkerOrder(MxS, Order=sample(MarkerNo(MxS))))
head(a)
head(b)
setequal(a,b)
}

