% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.R
\name{SetSampleFilter}
\alias{SetSampleFilter}
\title{Filtering the samples}
\usage{
SetSampleFilter(object, Filter = NULL, print.n.samples = T,
  step_name = "Filter samples")
}
\arguments{
\item{object}{\code{Genotype}. The object to be filtered.}

\item{Filter}{character, integer or logical. The samples to be filtered.
This could be a vector of sample names, sample indeces, or a logical vector
with TRUE values for samples to filter and FALSE for samples to keep.}

\item{print.n.samples}{logical. Should the number of remaining samples be
printed on the screen?}

\item{step_name}{character. The name of the filtration step. This will
in the step name column in the filter table.}
}
\value{
The filtered \code{Genotype} object. The filter table holds a new
line with the number of samples left after this step, and the
'Filtered_at_step' column in the SampleInfo of the filtered samples, have
the step name.
}
\description{
Filter (or 'mask') a specified set of samples from a \code{Genotype} object.
Update the filter table and the sample info.
}
\examples{
data("Mango_raw")
SampleNo(MxS)
MxS = SetSampleFilter(MxS, Filter=c("4_139", "77_13", "12"), step_name=
"arbitrary filtering")
getSampleInfo(MxS, Masked=F)["77_13", ]
}

