% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basics.R
\name{SetSampleOrder}
\alias{SetSampleOrder}
\title{Reorder the samples}
\usage{
SetSampleOrder(object, Order = NULL)
}
\arguments{
\item{object}{\code{Genotype}. The object to reorder the sample in}

\item{Order}{\code{numeric} or \code{character}. A numeric vector with the
order of samples, or a character vector with the sample names ordered as
desired.}
}
\value{
The object with the samples re-ordered as specified.
}
\description{
Reorder the samples
}
\examples{
data("Mango_raw")
a = SampleNames(MxS)
b = SampleNames(SetSampleOrder(MxS, Order=sample(SampleNo(MxS))))
head(a)
head(b)
setequal(a,b)
}

