% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_table.R
\name{start.filter.table}
\alias{start.filter.table}
\title{Creating a filter table for a \code{Genotype} object}
\usage{
start.filter.table(object, anyway = F)
}
\description{
This is a function which usually runs automatically with the first call to
SetMarkerFilter or to SetSampleFilter
}

