% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STRUCTURE_interface.R
\name{Sum_Structure_Results}
\alias{Sum_Structure_Results}
\title{Summing up multiple STRUCTUE simulations with predifined populations}
\usage{
Sum_Structure_Results(path, postfix = "_f", k, print_file = T)
}
\arguments{
\item{path}{\code{character}. The path to the directory where the STRUCTURE
output files are.}

\item{postfix}{\code{character}. The postfix of the file names. If STRUCTURE
run used \code{"write.structure.bin"}, then the file name is
\code{"K_<K>_run_<itteration number><postfix>"}.}

\item{k}{\code{numeric}. The number of sub-population or 'clusters'. This
number is a pre-defined parameter required by STRUCTURE.}

\item{print_file}{\code{logical}. Should the name of the each file read be
printed?}
}
\value{
the matrix of mean relative ancestry, summing over all simulations.
}
\description{
STRUCTURE calculates for each individual sample, how much of its genome
belongs to each one of K sub-populations, or so-called clusters.
\code{Sum_Structure_Results} sums up the results of a number of STRUCTURE
simulations, all ran with on the same data and with the same K. It can only
be used if STRCTURE used some samples with pre-defined populations.
Otherwise the 'phase' of the STRUCTURE output will be different in each run,
that is, population number 1 from one simulation, may have a different
number in another population. In this case, use tools like CLUMPP or CLUMPAK
which solve the phase problem first.
}
\seealso{
\code{\link{get_matrix}}
}

