% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{trio_genotype_match}
\alias{trio_genotype_match}
\title{Find the match of a potential parents-offspring trio}
\usage{
trio_genotype_match(voffspring, vparent1, vparent2)
}
\arguments{
\item{voffspring}{\code{character}. The genotype calls of the offspring, in
a "XX", "XY", "YY" format.}

\item{vparent1}{\code{character}. The genotype calls of the first parent, in
a "XX", "XY", "YY" format.}

\item{vparent1}{\code{character}. The genotype calls of the second parent,
in a "XX", "XY", "YY" format.}
}
\value{
\code{logical}. A vector specifying for each marker, if it supports
the possibility of the offspring being the offspring of the two parents.
}
\description{
Find the match of a potential parents-offspring trio
}

