% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Identical_samples.R
\name{Unique}
\alias{Unique}
\title{Remove redundency due to genetically identical samples}
\usage{
Unique(object, error.freq = 0)
}
\arguments{
\item{object}{\code{Genotype}. The name of the object to remove the
redundency from.}

\item{error.freq}{\code{numeric}. A single numeric value: the estimated
genotype call error frequency.}
}
\value{
\code{Genotype}. The filtered object.
}
\description{
This function filters the samples in a \code{Genotype} object, such that
from each group of genetically identical samples, only one representative
remains, and all the others are masked.
}
\details{
Any pair of samples, that have identical genotype call in
\code{1-error.freq} or more markers, are considered genetically identical.
\code{Unique} calls \code{FindIdentical} and \code{IdenticalGroups} to find
groups of identical samples. It chooses a sample with minimal number of
No-Calls as the of representative to keep from each group.
    After the filtration, the object slots 'SampleFilter' and 'filter_table'
are updated appropriately.
}
\examples{
data("Mango_row")
SampleNo(MxS)
MxS = Unique(MxS, error.freq=0.05)
}
\seealso{
\code{\link{FindIdentical}} for finding pairs of identical samples,
\code{\link{IdenticalGroups}} for creating groups of identical samples and
\code{\link{AddGroup2SampleInfo}} for adding the identity group to the
sample information.
}

