% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_table.R
\name{UpdateFilterTable}
\alias{UpdateFilterTable}
\title{Keeping the filter table updated}
\usage{
UpdateFilterTable(object, step_name)
}
\arguments{
\item{object}{\code{Genotype}. The object which has the table you need to
update.}

\item{step_name}{character. The name of the filtration step.}
}
\value{
The \code{Genotype} object with the new line in its filter table
specifying the step name and the number of markers/samples left after it.
}
\description{
the filter table keeps track of all markers and samples filtration steps.
each such step is followed by a call to \code{UpdateFilterTable}.
}

