% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cervus_interface.R
\name{write.candidates}
\alias{write.candidates}
\title{Write an input file with candidate parents for each offspring, for 'Cervus'
software}
\usage{
write.candidates(object, ...)
}
\arguments{
\item{object}{\code{Genotype}. The object to take the samples (offsprings
and candidate parents) from. Masked samples and markers will not be
included.}

\item{filename}{\code{character}. The file name for the candidate parents
file.}

\item{path}{\code{character}. The path for the candidate parents file.}

\item{identical_group}{\code{list}. A list of \code{character} vectors. Each
one is a set of sample names which are all genetically identical. Can be
created using the function \code{\link{IdenticalGroups}}.}

\item{ID_identical}{\code{IdenticalSamples}. A table of identical sample
pairs. Each row corresponds to one identical pair. The table must contain
columns called sample1 and sample2 with pair sample names. It can be created
using the function \code{\link{FindIdentical}}. No need to use this if
\code{identical_group} is present.}

\item{filter_by_generation}{\code{logical}. Should an information about the
generations of samples be used to eliminate candidate parents? If TRUE,
samples of younger generations cannot be candidate parents for sample from
older generations. To use this option, there must be a column of generation
information in the object's \code{SampleInfo} slot.}

\item{mm_cutoff}{\code{numeric}. The mis-match cutoff. A value between 0 and
1. The relative number of markers with no common allele that is allowed for
a candidate parent-offspring pair.}
}
\value{
\code{NULL}.
}
\description{
Cervus is a computer program for assignment of parents to their offspring
using genetic markers. It first calculates allele frequencies from a
genotype call file. Than it runs a simulation to determine the level of log
likelihood and delta required for parental assignment. Finally it assigns
uses a list of offsprings, and a list of candidate parents to each one of
them, to assign parents to offsprings, both when one of the parents is
known, and without a known parent.
}
\examples{
data("citrus_clean")
data("cit_id_groups")
write.candidates(MxS, identical_group=group, filter_by_generation=1,
mm_cutoff=0.95)

}
\seealso{
\code{\link{write.cervus}}, \code{\link{read.cervus}}
}

