% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cervus_interface.R
\name{write.cervus}
\alias{write.cervus}
\title{Write an input file with genotype calls for 'Cervus' software}
\usage{
write.cervus(object, fn, type = c("nuc", "numbers"))
}
\arguments{
\item{object}{\code{Genotype}. The object to use as input for Cervus. Must
be in "ACGT" format rather than in a "XY" format.
Masked samples and markers will not be included in this input.}

\item{fn}{\code{character}. The file name for the Cervus input file.}

\item{type}{\code{character}. The format (nuclotides or numbers) to use for
the Cervus input file.}
}
\value{
\code{NULL}.
}
\description{
Cervus is a computer program for assignment of parents to their offspring
using genetic markers. It first calculates allele frequencies from a
genotype call file. Than it runs a simulation to determine the level of log
likelihood and delta required for parental assignment. Finally it assigns
uses a list of offsprings, and a list of candidate parents to each one of
them, to assign parents to offsprings, both when one of the parents is
known, and without a known parent.
}
\examples{
data("citrus_clean")
Alleles = getMarkerInfo(MxS, Masked=FALSE)$Alleles
names(Alleles) <- MarkerNames(MxS, Masked=F)
nucMxS=xy2atgc(MxS, alleles=Alleles)
write.cervus(nucMxS, fn="citrus_genotype.txt", type="numbers")
}
\seealso{
\code{\link{xy2atgc}}, \code{\link{write.candidates}},
\code{\link{read.cervus}}
}

