% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_table.R
\name{write.filter.table}
\alias{write.filter.table}
\title{Printing the filter table to a csv file.}
\usage{
write.filter.table(object, path = "./", filename = "filter_table.csv")
}
\arguments{
\item{object}{\code{Genotype}. The object which has the table.}

\item{path}{\code{character}. The path to the folder to write the table in.}

\item{filenam}{\code{character}. the filename}
}
\value{
\code{NULL}
}
\description{
Printing the filter table to a csv file.
}

