% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STRUCTURE_interface.R
\name{write.structure.bin}
\alias{write.structure.bin}
\title{writing an executable file to run the 'structure' commands}
\usage{
write.structure.bin(object, fn = "structure_sim", k.range = c(1, 10),
  n.runs = 20, path = "Structure", parameter_list = list())
}
\arguments{
\item{object}{\code{Genotype}. The object containing the samples and markers
to run STRUCTURE on.}

\item{fn}{\code{character}. The file name for the file with the STRUCTURE commands.}

\item{k.range}{\code{numeric}. A vector with 2 values: the minimum and
maximum 'K's to run. The program will run all the values of K between and
including the minimum and maximum.}

\item{n.runs}{\code{numeric}. The number of runs to run each K. In any case,
there will not be more than one run with K=1, because for this K the results
are degenerated and deterministic.}

\item{path}{\code{character}. The directory to write the STRUCTURE files in.}

\item{parameter_list}{\code{list}. An optional list of additional parameters
for STRUCTURE.}
}
\description{
Writing an 'bash' file with the command to run a series of STRUCTURE
simulations, and files with parameters for STRUCTURE. This allows you to run
a number of simulations for each K (the number of ancestry populations),
which is recommended for the analysis of STRUCTURE results.
}
\details{
After running \code{write.structure.bin}, a new file named
\code{fn} is formed in your working directory, with the calls to STRUCTURE.
to run it, you first need to turn it to an executable file, for example by
typing \code{chmod a=rwx <fn>} in your Unix command line terminal.  You can
then run it by typing in your Unix command line terminal
    \code{./<fn>}.
}
\examples{
data("citrus_clean")
Alleles = getMarkerInfo(MxS, Masked=FALSE)$Alleles
names(Alleles) <- MarkerNames(MxS, Masked=F)
nucMxS=xy2atgc(MxS, alleles=Alleles)
write.structure.bin(MxS, fn="structure_sim", k.range = c(1, 10), n.runs=20)
}

