% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STRUCTURE_interface.R
\name{write.structure}
\alias{write.structure}
\title{Write an input file for 'STRUCTURE' software}
\usage{
write.structure(object, fn, type = c("nuc", "numbers"), popdata = NULL)
}
\arguments{
\item{object}{\code{Genotype}. The object to use as input for STRUCTURE.
Masked samples and markers will not be included in this input.}

\item{fn}{\code{character}. The file name for the STRUCTURE input file.}

\item{type}{\code{character}. The format (nuclotides or numbers) to use for
the STRUCTURE input file.}

\item{popdata}{\code{numeric}. An optional argument. A named numeric vector.
The population each sample belongs to. Use 0 for samples of unknown
population. The names are the names of samples in \code{object}.}
}
\value{
\code{NULL}.
}
\description{
STRUCTURE is a free software package for using multi-locus genotype data to
investigate population structure. It infers the presence of distinct
populations, assigns individuals to populations, identifies migrants and
admixed individuals, and estimates population allele frequencies. It is
therefore desireable for your project to write an input for it, and read its
output.
\code{write.structure} writes an input file for STRUCTURE.
}
\examples{
data("citrus_clean")
Alleles = getMarkerInfo(MxS, Masked=FALSE)$Alleles
names(Alleles) <- MarkerNames(MxS, Masked=F)
nucMxS=xy2atgc(MxS, alleles=Alleles)
write.structure(nucMxS, fn="Structure_input.txt", type="numbers")
}

