% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBD.R
\name{write_snpgds_file}
\alias{write_snpgds_file}
\title{Write a \code{SNPRelate} package \code{snpgds} file from \code{Genotype}}
\usage{
write_snpgds_file(Genotype, filename = "Genotype.gds")
}
\arguments{
\item{filename}{\code{character}. The name of the snpgds file.}

\item{object}{\code{Genotype}. The object to write to a gds file.}
}
\value{
\code{character}. The name of the file created.
}
\description{
Changing from a \code{Genotype} object to a gds file allows us to use
\code{gdsfmt} package functions
}
\examples{
data("Genotype_43X80")
write_snpgds_file(Genotype, filename="Genotype.gds")
}

