% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STRUCTURE_interface.R
\name{xy2atgc}
\alias{xy2atgc}
\title{Change genotype call allele format from call from X and Y to A, C, T and G}
\usage{
xy2atgc(object, alleles = character())
}
\arguments{
\item{object}{\code{Genotype}. The object to change the genotype call
encoding in.}

\item{alleles}{\code{character}. A named character vector of length
\code{MakrerNo(object, Masked=F)} with the 2 alleles of each marker. For a
marker with the two alleles, "A" and "G" the value will be "AG". The names
of the values are the marker names, which should be the same as the
\code{Genotype} object mareker names.}
}
\value{
\code{Genotype}. The same object but with genotype calls specified
in nucleotides.
}
\description{
In order to use some downstream tools, it is sometimes required to transorm
the coding of the genotype calls from "XX", "XY" and "YY" to nucleotide
based, like "CC", "GT" and so on.
}
\examples{
data("citrus_clean")
Alleles = getMarkerInfo(MxS, Masked=FALSE)$Alleles
names(Alleles) <- MarkerNames(MxS, Masked=F)
nucMxS=xy2atgc(MxS, alleles=Alleles)
}

