% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBD.R
\name{xy2gds}
\alias{xy2gds}
\title{Write a \code{gdsfmt} package gds file from \code{Genotype}}
\usage{
xy2gds(object, gds.fn = "SNP.gds", Chrom = NULL, Pos = NULL, Allele)
}
\arguments{
\item{object}{\code{Genotype}. The object to write to a gds file.}

\item{gds.fn}{\code{character}. The name of the gds file.}

\item{Chrom}{\code{character}. Chromosome names markers.}

\item{Pos}{\code{numeric}. Chromosomal positions markers.}

\item{Allele}{\code{character}. Marker alleles.}
}
\value{
\code{logical}
}
\description{
Changing from a \code{Genotype} object to a gds file allows us to use
\code{gdsfmt} package functions
}

